+++ packages/core/src/utils/handle-class-name.native.ts
@@ -1,6 +1,6 @@
 import { Variants } from "@/types";
 
 export const nativeHandleClassName = (variant: Variants, className: string): string =>
-  `${
+  `icon-default ${
     (variant === "filled" && "icon-filled ") || (variant === "outlined" && "icon-outlined ") || ""
-  }${((!className || !/icon-/.exec(className)) && "icon-default ") || ""}${className}`.trim();
+  }${className ?? ""}`.trim();