+++ packages/tailwind/src/utils/types.ts
@@ -1,10 +1,3 @@
-import { Style } from "@/types";
-
-export interface IStyleGenerator {
-  add(name: string, styles: string): void;
-  styles: Style;
-}
-
 export type SpecialStyleProps = {
   h: string;
   w: string;