+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+## [0.3.4](https://github.com/rocketclimb/rocketicons/compare/v0.3.3...v0.3.4) (2024-05-25)
+
+### Bug Fixes
+
+- **icons** improve class name handling on native [[core](./packages/core/CHANGELOG.md), [tailwind](./packages/tailwind/CHANGELOG.md)] ([6575100f](https://github.com/rocketclimb/rocketicons/commit/6575100fa60dd11dd3086225d25ebc532805cda2))
+
+### Other
+
+-  ([ffdc4f63](https://github.com/rocketclimb/rocketicons/commit/ffdc4f63d82267c407aa8bb1173917eea9fc6922))
+
+## [0.3.4](https://github.com/rocketclimb/rocketicons/compare/v0.3.3...v0.3.4) (2024-05-25)
+
+### Bug Fixes
+
+- **icons** improve class name handling on native [[core](./packages/core/CHANGELOG.md), [tailwind](./packages/tailwind/CHANGELOG.md)] ([6575100f](https://github.com/rocketclimb/rocketicons/commit/6575100fa60dd11dd3086225d25ebc532805cda2))
+
+### Other
+
+-  ([ffdc4f63](https://github.com/rocketclimb/rocketicons/commit/ffdc4f63d82267c407aa8bb1173917eea9fc6922))
+
+## [0.3.4](https://github.com/rocketclimb/rocketicons/compare/v0.3.3...v0.3.4) (2024-05-25)
+
+### Bug Fixes
+
+- **icons** improve class name handling on native [[core](./packages/core/CHANGELOG.md), [tailwind](./packages/tailwind/CHANGELOG.md)] ([6575100f](https://github.com/rocketclimb/rocketicons/commit/6575100fa60dd11dd3086225d25ebc532805cda2))
+
+## [0.3.4](https://github.com/rocketclimb/rocketicons/compare/v0.3.3...v0.3.4) (2024-05-25)
+
+### Bug Fixes
+
+- **icons** improve class name handling on native [[core](./packages/core/CHANGELOG.md), [tailwind](./packages/tailwind/CHANGELOG.md)] ([6575100f](https://github.com/rocketclimb/rocketicons/commit/6575100fa60dd11dd3086225d25ebc532805cda2))
+
 ## [0.3.3](https://github.com/rocketclimb/rocketicons/compare/v0.3.2-release...v0.3.3) (2024-05-22)
 
 ### Bug Fixes
@@ -32,6 +64,7 @@
 - playground url fixed [[ignition](./packages/ignition/CHANGELOG.md)] ([8355e217](https://github.com/rocketclimb/rocketicons/commit/8355e217ca22d848d6bd22a07bd7b8f49378a8b2))
 
 ## [0.3.1](https://github.com/rocketclimb/rocketicons/compare/v0.3.0-release...v0.3.1) (2024-05-17)
+
 - landingpage layout fixes
 
 ## [0.3.0](https://github.com/rocketclimb/rocketicons/compare/v0.3.0-release...v0.3.1) (2024-05-17)