+++ packages/ignition/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.3.5](https://github.com/rocketclimb/rocketicons/compare/v0.3.3...v0.3.4) (2024-05-25)
+
+### Other
+
+-  ([ffdc4f63](https://github.com/rocketclimb/rocketicons/commit/ffdc4f63d82267c407aa8bb1173917eea9fc6922))
+
 ## [0.3.4](https://github.com/rocketclimb/rocketicons/compare/v0.3.2-release...v0.3.3) (2024-05-22)
 
 ### Bug Fixes